Unify project and add units
- The data logger now takes in a string of data sent by the robot once it
is disabled. The string is comma separated numbers which are then
transformed into a 2d list containing all the sets of collected data.
- The drivetrain template has been reduced from 4 different templates to a
single template and accomodate for the other tests that only require 1
set of motor controllers.
- This template also concatenates gathered data values into a string and
sends it to the data logger after the run is over allowing us to remove
the network tables calls during autonomous that could not be refreshed
quickly. Also the dt of the robot notifier has been updated to .01s for
a faster update rate.
- There is now a single data analyzer to take care of all the different
tests
- The logger outputs units and test type to facilitate analyzing and unit
conversion
-Added the option to convert PID gains into units to run in native
controller methods (e.g. Spark Max and Talon SRX).
- Replace wheel diameter with units per rotation