Fixed concatenate error
For some reason that eludes me the order of variables passed into concatenate causes collected data to not be properly concatenated https://imgur.com/vijWbFG
I did a few tests and no data is lost, it just is not properly placed. I am unsure as to the reason why but moving the variables to this order solved the problem that has been plaguing my team for multiple weeks. https://imgur.com/rBg6WMd